+++ stories/components/progress-radial/progress-radial.stories.ts
@@ -1,4 +1,4 @@
-import { Meta, StoryFn } from '@storybook/web-components';
+import { Meta, StoryFn, StoryObj } from '@storybook/web-components';
 import { html } from 'lit-html';
 import { getComponentArgs } from '../../utils/args-generation';
 
@@ -9,22 +9,32 @@ const meta: Meta = {
 };
 export default meta;
 
-const Template: StoryFn<Partial<{size: number, showLabel: boolean, value: number}>> = (args) => html`
+type ProgressRadialArgs = {
+    size: number;
+    showLabel: boolean;
+    value: number;
+}
+
+const Template: StoryFn<Partial<ProgressRadialArgs>> = (args) => html`
     <wcs-progress-radial .size=${args.size}
                          ?show-label=${args.showLabel}
                          .value=${args.value}>
     </wcs-progress-radial>
 `;
 
-export const Default = Template.bind({});
-Default.args = {
-    size: 120,
-    value: 25
+export const Default: StoryObj = {
+    render: (args: ProgressRadialArgs) => Template(args, this),
+    args: {
+        size: 120,
+        value: 25
+    },
 };
 
-export const WithLabel = Template.bind({});
-WithLabel.args = {
-    size: 120,
-    value: 25,
-    showLabel: true
-};
+export const WithLabel: StoryObj = {
+    render: (args: ProgressRadialArgs) => Template(args, this),
+    args: {
+        size: 120,
+        value: 25,
+        showLabel: true
+    }
+}